meson: Attempt to use link_depends to re-link libOpenGL when gen_libOpenGL_exports.py changes
This doesn't quite work: meson generates an order-only dependency on
src/generate/g_opengl_exports.sym for src/OpenGL/OpenGL@sha/libopengl.c.o, but
no actual dependency on the link step for src/OpenGL/libOpenGL.so.0.0.0.

Also, it would be nice to put this link_depends line in g_opengl_exports_sym_dep
rather than having to have it in the shared_library() call, but that's a minor
detail.